// frontend/src/api/rest/api.ts
export interface AddTrackUrlRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
    /**
     * Ссылка на трек
     */
    'trackUrl': string;
}
/**
 * Ответ на запрос добавления в заказ ссылки на трек
 */
export interface AddTrackUrlResponse {
    /**
     * Заказ разбора трека
     */
    'reviewOrder': ReviewOrderDto;
}
/**
 * Настройки приложения
 */
export interface AppSettingsDto {
    /**
     * Номинальная стоимость заказа
     */
    'reviewOrderNominalAmount': number;
}
/**
 * Запрос отмены заказа
 */
export interface CancelReviewOrderRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
}
// ...
export interface CompleteReviewOrderRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
    /**
     * Оценка трека (0-26)
     */
    'rating': number;
}
// ...
export interface ComposerStreamDto {
    /**
     * Id стрима
     */
    'id': number;
    /**
     * Дата проведения стрима
     */
    'eventDate': string;
    /**
     * Статус стрима
     */
    'status': ComposerStreamStatus;
    /**
     * Тип стрима
     */
    'type': ComposerStreamType;
    /**
     * Дата и время начала стрима
     */
    'startedAt': string | null;
    /**
     * Дата и время завершения стрима
     */
    'completedAt': string | null;
}
// ...
export interface CreateReviewOrderRequest {
    /**
     * Псевдоним пользователя
     */
    'nickname': string;
    /**
     * Тип заказа разбора трека
     */
    'orderType': ReviewOrderType;
    /**
     * Ссылка на трек
     */
    'trackUrl'?: string | null;
    /**
     * Сумма платежа
     */
    'paymentAmount'?: number | null;
    /**
     * Комментарий пользователя
     */
    'userComment'?: string | null;
}
// ...
export interface FreezeReviewOrderRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
}
// ...
export interface MoveUpReviewOrderRequest {
    /**
     * Псевдоним пользователя
     */
    'nickname': string;
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
    /**
     * Сумма платежа
     */
    'paymentAmount': number;
}
/**
 * Ответ на запрос поднятия заказа в очереди
 */
export interface MoveUpReviewOrderResponse {
    /**
     * Заказ разбора трека
     */
    'reviewOrder': ReviewOrderDto;
    /**
     * Id платежа
     */
    'paymentTransactionId': number;
}

export const OrderActivityStatus = {
    Unspecified: 'Unspecified',
    Active: 'Active',
    InProgress: 'InProgress',
    Completed: 'Completed',
    Scheduled: 'Scheduled',
    Frozen: 'Frozen',
    Removed: 'Removed'
} as const;

export type OrderActivityStatus = typeof OrderActivityStatus[keyof typeof OrderActivityStatus];


/**
 * Тип категории заказа
 */

export const OrderCategoryType = {
    Unspecified: 'Unspecified',
    OutOfQueue: 'OutOfQueue',
    Donation: 'Donation',
    Debt: 'Debt'
} as const;

export type OrderCategoryType = typeof OrderCategoryType[keyof typeof OrderCategoryType];


/**
 * Представляет позицию заказа в очереди, включая сам заказ и историю перемещений
 */
export interface OrderPositionDto {
    /**
     * Заказ разбора трека
     */
    'order': ReviewOrderDto;
    /**
     * Предыдущая позиция заказа в очереди
     */
    'previousPosition': OrderQueuePositionDto;
    /**
     * Текущая позиция заказа в очереди
     */
    'currentPosition': OrderQueuePositionDto;
}
/**
 * Позиция заказа в очереди, включая его индекс, статус активности и категорию
 */
export interface OrderQueuePositionDto {
    /**
     * Позиция заказа в очереди
     */
    'queueIndex': number;
    /**
     * Статус активности заказа
     */
    'activityStatus': OrderActivityStatus;
    /**
     * Тип категории заказа
     */
    'categoryType': OrderCategoryType;
    /**
     * Индекс категории, если заказ относится к долговой категории
     */
    'categoryDebtIndex': number;
}


/**
 * Запрос регистрации пользователя
 */
export interface RegisterRequest {
    /**
     * Имя пользователя
     */
    'userName': string;
    /**
     * Пароль
     */
    'password': string;
}
/**
 * Заказ разбора трека
 */
export interface ReviewOrderDto {
    /**
     * Id заказа
     */
    'id': number;
    /**
     * Дата и время создания заказа
     */
    'createdAt': string;
    /**
     * Дата и время взятия заказа в работу
     */
    'inProgressAt': string | null;
    /**
     * Дата и время выполнения заказа
     */
    'completedAt': string | null;
    /**
     * Тип заказа разбора трека
     */
    'type': ReviewOrderType;
    /**
     * Тип категории заказа
     */
    'categoryType': OrderCategoryType;
    /**
     * Статус заказа разбора трека
     */
    'status': ReviewOrderStatus;
    /**
     * Заказ заморожен
     */
    'isFrozen': boolean;
    /**
     * Ссылка на трек
     */
    'trackUrl': string | null;
    /**
     * Комментарий пользователя
     */
    'userComment': string | null;
    /**
     * Основной ник пользователя, из всех пользователей, кто причастен к созданию заказа
     */
    'mainNickname': string;
    /**
     * Общая стоимость заказа (номинал + платежи)
     */
    'totalAmount': number;
    /**
     * Связанный cтрим композитора, где создан заказ
     */
    'creationStream': ComposerStreamDto;
}


/**
 * Статус заказа разбора трека
 */

export const ReviewOrderStatus = {
    Unspecified: 'Unspecified',
    Preorder: 'Preorder',
    Pending: 'Pending',
    InProgress: 'InProgress',
    Completed: 'Completed',
    Canceled: 'Canceled'
} as const;

export type ReviewOrderStatus = typeof ReviewOrderStatus[keyof typeof ReviewOrderStatus];


/**
 * Тип заказа разбора трека
 */

export const ReviewOrderType = {
    Unspecified: 'Unspecified',
    OutOfQueue: 'OutOfQueue',
    Donation: 'Donation',
    Free: 'Free',
    Charity: 'Charity',
    Custom: 'Custom'
} as const;

export type ReviewOrderType = typeof ReviewOrderType[keyof typeof ReviewOrderType];


/**
 * Запрос запуска стрима
 */
export interface StartStreamRequest {
    /**
     * Id стрима
     */
    'composerStreamId': number;
}
/**
 * Ответ на запрос запуска стрима
 */
export interface StartStreamResponse {
    /**
     * Стрим композитора
     */
    'composerStream': ComposerStreamDto;
}
/**
 * Запрос взятия заказа в работу
 */
export interface TakeOrderInProgressRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
}
/**
 * Ответ на запрос взятия заказа в работу
 */
export interface TakeOrderInProgressResponse {
    /**
     * Заказ разбора трека
     */
    'reviewOrder': ReviewOrderDto;
}
/**
 * Запрос разморозки заказа
 */
export interface UnfreezeReviewOrderRequest {
    /**
     * Id заказа разбора трека
     */
    'reviewOrderId': number;
}
// ...
export class OrderQueueApi extends BaseAPI {
    /**
     * 
     * @summary Получает текущее состояние очереди заказов
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiOrderQueueGetOrderQueueGet(options?: RawAxiosRequestConfig) {
        return OrderQueueApiFp(this.configuration).apiOrderQueueGetOrderQueueGet(options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ReviewOrderApi extends BaseAPI {
    /**
     * 
     * @summary Добавляет или изменяет ссылку на трек
     * @param {AddTrackUrlRequest} [addTrackUrlRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderAddTrackUrlPost(addTrackUrlRequest?: AddTrackUrlRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderAddTrackUrlPost(addTrackUrlRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Отменяет заказ
     * @param {CancelReviewOrderRequest} [cancelReviewOrderRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderCancelReviewOrderPost(cancelReviewOrderRequest?: CancelReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderCancelReviewOrderPost(cancelReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Выполнение заказа
     * @param {CompleteReviewOrderRequest} [completeReviewOrderRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderCompleteReviewOrderPost(completeReviewOrderRequest?: CompleteReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderCompleteReviewOrderPost(completeReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Создает заказ
     * @param {string} [idempotencyKey] Ключ идемпотентности
     * @param {CreateReviewOrderRequest} [createReviewOrderRequest] Запрос создания заказа
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderCreateReviewOrderPost(idempotencyKey?: string, createReviewOrderRequest?: CreateReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderCreateReviewOrderPost(idempotencyKey, createReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Замораживает заказ
     * @param {FreezeReviewOrderRequest} [freezeReviewOrderRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderFreezeReviewOrderPost(freezeReviewOrderRequest?: FreezeReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderFreezeReviewOrderPost(freezeReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Поднимает заказ в очереди
     * @param {string} [idempotencyKey] Ключ идемпотентности
     * @param {MoveUpReviewOrderRequest} [moveUpReviewOrderRequest] Запрос поднятия заказа в очереди
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderMoveUpReviewOrderPost(idempotencyKey?: string, moveUpReviewOrderRequest?: MoveUpReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderMoveUpReviewOrderPost(idempotencyKey, moveUpReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Взятие заказа в работу
     * @param {TakeOrderInProgressRequest} [takeOrderInProgressRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderTakeOrderInProgressPost(takeOrderInProgressRequest?: TakeOrderInProgressRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderTakeOrderInProgressPost(takeOrderInProgressRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Размораживает заказ
     * @param {UnfreezeReviewOrderRequest} [unfreezeReviewOrderRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiReviewOrderUnfreezeReviewOrderPost(unfreezeReviewOrderRequest?: UnfreezeReviewOrderRequest, options?: RawAxiosRequestConfig) {
        return ReviewOrderApiFp(this.configuration).apiReviewOrderUnfreezeReviewOrderPost(unfreezeReviewOrderRequest, options).then((request) => request(this.axios, this.basePath));
    }
}